+++ firestore-palm-chatbot/functions/__tests__/generative/index.test.ts
@@ -22,7 +22,7 @@ jest.mock('../../src/config', () => ({
     enableDiscussionOptionOverrides: true,
     candidatesField: 'candidates',
     provider: 'generative',
-    model: 'chat-bison-001',
+    model: 'chat-bison',
   },
 }));
 
@@ -281,7 +281,7 @@ describe('generateMessage', () => {
     // verify SDK is called with expected arguments
     const expectedRequestData = {
       candidateCount: undefined,
-      model: 'models/chat-bison-001',
+      model: 'models/chat-bison',
       prompt: {
         messages: [
           {
@@ -371,7 +371,7 @@ describe('generateMessage', () => {
     // verify SDK is called with expected arguments
     const expectedRequestData = {
       candidateCount: undefined,
-      model: 'models/chat-bison-001',
+      model: 'models/chat-bison',
       prompt: {
         messages: [
           {